# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Pershing307 Member Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/domains:
    get:
      operationId: getUserDomains
      tags: [users]
      summary: List domain access grants for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain access list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDomainAccessListResult"
    post:
      operationId: grantUserDomain
      tags: [users]
      summary: Grant a protected domain access to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainGrantInput"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/domains/{domainId}:
    delete:
      operationId: revokeUserDomain
      tags: [users]
      summary: Revoke protected domain access from a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Access revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List all invitations (admin)
      responses:
        "200":
          description: Invitation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListResult"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Create a new invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreatedResult"
        "409":
          description: User or invitation already exists
  /invitations/accept/{token}:
    get:
      operationId: getInvitationByToken
      tags: [invitations]
      summary: Look up invitation by token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationLookupResult"
        "404":
          description: Invalid or expired invitation
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation and create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "404":
          description: Invalid or expired invitation
  /invitations/{id}:
    delete:
      operationId: revokeInvitation
      tags: [invitations]
      summary: Revoke a pending invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
        "404":
          description: Invitation not found
  /invitations/{id}/link:
    get:
      operationId: getInvitationLink
      tags: [invitations]
      summary: Get the invitation accept link (for copy when SMTP is unavailable)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationLinkResult"
        "404":
          description: Invitation not found or already accepted
...
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    UserDomainGrant:
      type: object
      required: [domainId, domainName, domainSlug, grantedAt]
      properties:
        domainId:
          type: string
        domainName:
          type: string
        domainSlug:
          type: string
        grantedAt:
          type: string
    UserDomainAccessListResult:
      type: object
      required: [domains]
      properties:
        domains:
          type: array
          items:
            $ref: "#/components/schemas/UserDomainGrant"
    DomainGrantInput:
      type: object
      required: [domainId]
      properties:
        domainId:
          type: string
          minLength: 1
...
    InvitationSummary:
      type: object
      required: [id, email, firstName, lastName, expiresAt, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        expiresAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        revokedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    InvitationListResult:
      type: object
      required: [invitations, smtpConfigured]
      properties:
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/InvitationSummary"
        smtpConfigured:
          type: boolean
    InvitationInput:
      type: object
      required: [email, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        roleId:
          type: ["string", "null"]
    InvitationCreated:
      type: object
      required: [id, email, firstName, lastName, expiresAt]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        expiresAt:
          type: string
    InvitationCreatedResult:
      type: object
      required: [invitation, smtpConfigured]
      properties:
        invitation:
          $ref: "#/components/schemas/InvitationCreated"
        smtpConfigured:
          type: boolean
    InvitationLookup:
      type: object
      required: [email, firstName, lastName, lodgeName]
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        lodgeName:
          type: string
    InvitationLookupResult:
      type: object
      required: [invitation]
      properties:
        invitation:
          $ref: "#/components/schemas/InvitationLookup"
    AcceptInvitationInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
    InvitationLinkResult:
      type: object
      required: [link]
      properties:
        link:
          type: string
...