# lib/api-zod/src/generated/types/invitationListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */
import type { InvitationSummary } from './invitationSummary';

export interface InvitationListResult {
  invitations: InvitationSummary[];
  smtpConfigured: boolean;
}