# lib/api-zod/src/generated/types/userDomainAccessListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */
import type { UserDomainGrant } from './userDomainGrant';

export interface UserDomainAccessListResult {
  domains: UserDomainGrant[];
}